+++ docs/Changelog.md
@@ -27,6 +27,7 @@
 * [#7165](https://github.com/TouK/nussknacker/pull/7165) Added PerformSingleExecution scenario action
     * Added support for PerformSingleExecution action in DeploymentManager and in GUI
     * Improved scenario state management to include information about current and deployed versions and allow more customization
+* [#7184](https://github.com/TouK/nussknacker/pull/7184) Improve Nu Designer API notifications endpoint, to include events related to currently displayed scenario
 
 ### 1.18.1 (9 December 2024)